+++ polars-cv/tests/test_schema_inference.py
@@ -200,7 +200,7 @@ class TestNullDataSchemaConsistency:
         df = pl.DataFrame({"path": pl.Series("path", [None], dtype=pl.String)})
         pipe = (
             Pipeline()
-            .source("file_path")
+            .source("file_path", dtype="u8")
             .assert_shape(height=10, width=10, channels=3)
             .sink("array")
         )
